Flatten handling of passwords (environment vs. files)
If the value of BANPROXY_PASSWORD is a file, the contents of the
file will be used as the parameter banproxy.password. If the value
of the variable is not a file, its value will be used as the
password.

The change is largely designed to make it easier to move between
different orchestrators with a single image.